requirements.txt
# [FastAPI Core] 웹 서버 및 데이터 검증
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
python-multipart>=0.0.9    # 파일 업로드 (사진 처리용)

# [Azure SDK] 클라우드 인프라 연동
azure-cosmos>=4.5.1        # Cosmos DB 세션 저장용
azure-storage-blob>=12.19.0 # 3D/영상 파일 서빙 및 SAS 토큰 발급용
azure-search-documents>=11.4.0 # AI Search (RAG 검색용)

# [AI & Vision] 모델 및 데이터 처리
openai>=1.14.0             # GPT-4o 호출용 (Azure OpenAI 호환)
mediapipe>=0.10.11         # Scene 2, 3 체형(관절) 분석용
opencv-python-headless>=4.9.0 # 이미지 처리용 (서버용이라 headless 사용)

# [Utilities] 기타 도구
python-dotenv>=1.0.1       # .env 파일에서 환경변수 읽기